requirements.txt
python-telegram-bot==20.7
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0
httpx==0.26.0
anthropic==0.18.1
